Testing: Cross check text and json mode
Merge FatraceRunner* classes into one.

Let class FatraceRunner run two instances of fatrace in parallel, one
with and one without --json. Each assert provides a regex for testing
the text mode log and a lambda for testing the JSON mode log.

The regex and lambda should match the same sets of events. However, we
can't rely on the assumption that the two fatrace instances catch the
exact same sequence of events. Instead, for each assert and text mode
output line, we assert that the regex matches the line iff the lambda
matches it after parsing.